agent/requirements.txt
python-dotenv==1.1.1
requests==2.32.3
pandas==2.3.2
pyodbc==5.2.0
geopandas==1.1.1
shapely==2.1.1
